# Conflicts:
#	VisionpredatorProcess/BulletSystem.cpp
#	VisionpredatorProcess/BulletSystem.h
#	VisionpredatorProcess/DataRegister_Vispred.cpp

+++ VPGraphics/vpGraphics.h
@@ -137,7 +137,7 @@ namespace ui
 		float PosYPercent = 0.f;
 		float Scale = 1.f;
 		float Angle = 0.f;
-		uint32_t Layer = 0;
+		float Layer = 0;
 
 		// effect
 		bool isBold = false;
